requirements.txt
librosa
numpy
opencc_python_reimplemented
pydub
PyYAML
soundfile 
torch
torchvision 
torchaudio
git+https://github.com/openai/whisper.git